pyproject.toml
[project]
name = "undead-slayer"
version = "0.1.0"
description = "An RPG game."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
    "rich-pixels>=3.0.1",
    "rich>=13.8.1",
    "textual>=0.81.0",
    "typer>=0.12.5",
    "xdg-base-dirs>=6.0.1",
]

[project.scripts]
undead_slayer = "undead_slayer.__main__:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "babel>=2.16.0",
    "pre-commit>=3.8.0",
    "pyright[nodejs]>=1.1.382.post1",
    "pytest>=8.3.3",
    "ruff>=0.6.8",
    "textual-dev>=1.6.1",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.pyright]
include = [
    "src/**/*.py"
    ]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/build",
    ]
ignore = [
    "**/.venv"
    ]
venvPath = "."
venv = ".venv"
typeCheckingMode = "strict"
